refactor(cc-logs-instances): move story fixtures to appropriate directory
